Revert "Move generation of annotations to `autosave`, because `save` doesn't get called on corrections"
This reverts commit 89fb285283232cbd243fc01659122380963d6529.
